pyproject.toml
[project]
name = "emoji-list"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "jijinbei", email = "hikoharu14142@outlook.jp" }
]
dependencies = [
    "openai>=1.58.1",
    "watchdog>=6.0.0",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/emoji_list"]

[tool.rye.scripts]
emoji-list = { cmd = "python ./src/main.py" }
